@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import List, TYPE_CHECKING, Optional
 
-from pydantic import BaseModel, parse_obj_as
+from pydantic import BaseModel, ConfigDict
 
 from prodsys.simulation import sim, store
 
@@ -25,8 +25,7 @@ class QueueFactory(BaseModel):
 
     queues: List[store.Queue] = []
 
-    class Config:
-        arbitrary_types_allowed = True
+    model_config = ConfigDict(arbitrary_types_allowed=True)
 
     def create_queues(self, adapter: adapter.ProductionSystemAdapter):
         """
@@ -66,59 +65,3 @@ class QueueFactory(BaseModel):
             List[store.Queue]: List of queue objects with the given IDs.
         """
         return [q for q in self.queues if q.queue_data.ID in IDs]
-    
-
-
-class StorageFactory(BaseModel):
-    """
-    Factory class that creates and stores `prodsys.simulation` queue objects from `prodsys.models` queue objects.
-
-    Args:
-        env (sim.Environment): prodsys simulation environment.
-
-
-    Returns:
-        _type_: _description_
-    """
-    env: sim.Environment
-
-    storages: List[store.Queue] = []
-
-    class Config:
-        arbitrary_types_allowed = True
-
-    def create_storages(self, adapter: adapter.ProductionSystemAdapter):
-        """
-        Creates queue objects based on the given adapter.
-
-        Args:
-            adapter (adapter.ProductionSystemAdapter): _description_
-        """
-        for data in adapter.storage_data:
-            self.add_storage(data)
-
-    def add_storage(self, storage_data: queue_data.StorageData):
-        values = {}
-        values.update({"env": self.env, "data": storage_data})
-        storage_object = parse_obj_as(store.Storage, values)
-
-        self.storages.append(storage_object)
-    
-
-    def get_storage(self, ID: str) -> Optional[store.Storage]:
-        """
-        Returns a process object based on the given ID.
-
-        Args:
-            ID (str): ID of the process object.
-
-        Raises:
-            ValueError: If the process object is not found.
-
-        Returns:
-            Optional[process.PROCESS_UNION]: Process object based on the given ID.
-        """
-        pr = [pr for pr in self.storages if pr.data.ID in ID]
-        if not pr:
-            raise ValueError(f"Process with ID {ID} not found")
-        return pr.pop()
