@@ -1,11 +1,6 @@
-from __future__ import annotations
+from typing import Union, Optional, Generator
 
-from abc import ABC
-from enum import Enum
-from collections.abc import Iterable
-from typing import List, Union, Optional, TYPE_CHECKING, Generator
-
-from pydantic import BaseModel, Field, Extra
+from pydantic import BaseModel, ConfigDict, Field
 
 import logging
 logger = logging.getLogger(__name__)
@@ -13,24 +8,23 @@ logger = logging.getLogger(__name__)
 import numpy as np
 from simpy import events
 
+from prodsys.models import product_data
+
 from prodsys.simulation import (
+    request,
     auxiliary,
     process,
-    request,
     router,
     resources,
     sim,
     sink,
     source,
     proces_models,
-    state,
     node
 )
+from prodsys.simulation.state import StateTypeEnum, StateEnum
 
-from prodsys.models import product_data
-
-
-class ProductInfo(BaseModel, extra=Extra.allow):
+class ProductInfo(BaseModel):
     """
     Class that represents information of the current state of a product.
 
@@ -46,14 +40,16 @@ class ProductInfo(BaseModel, extra=Extra.allow):
     resource_ID: str = Field(init=False, default=None)
     state_ID: str = Field(init=False, default=None)
     event_time: float = Field(init=False, default=None)
-    activity: state.StateEnum = Field(init=False, default=None)
+    activity: StateEnum = Field(init=False, default=None)
     product_ID: str = Field(init=False, default=None)
-    state_type: state.StateTypeEnum = Field(init=False, default=None)
+    state_type: StateTypeEnum = Field(init=False, default=None)
+
+    model_config=ConfigDict(extra="allow")
 
     def log_finish_product(
         self,
         resource: Union[resources.Resource, sink.Sink, source.Source],
-        _product: Product,
+        _product: "Product",
         event_time: float,
     ):
         """
@@ -68,13 +64,13 @@ class ProductInfo(BaseModel, extra=Extra.allow):
         self.state_ID = resource.data.ID
         self.event_time = event_time
         self.product_ID = _product.product_data.ID
-        self.activity = state.StateEnum.finished_product
-        self.state_type = state.StateTypeEnum.sink
+        self.activity = StateEnum.finished_product
+        self.state_type = StateTypeEnum.sink
 
     def log_create_product(
         self,
         resource: Union[resources.Resource, sink.Sink, source.Source],
-        _product: Product,
+        _product: "Product",
         event_time: float,
     ) -> None:
         """
@@ -89,15 +85,15 @@ class ProductInfo(BaseModel, extra=Extra.allow):
         self.state_ID = resource.data.ID
         self.event_time = event_time
         self.product_ID = _product.product_data.ID
-        self.activity = state.StateEnum.created_product
-        self.state_type = state.StateTypeEnum.source
+        self.activity = StateEnum.created_product
+        self.state_type = StateTypeEnum.source
 
     def log_start_process(
         self,
         resource: resources.Resource,
-        _product: Product,
+        _product: "Product",
         event_time: float,
-        state_type: state.StateTypeEnum,
+        state_type: StateTypeEnum,
     ) -> None:
         """
         Logs the start of a process.
@@ -112,15 +108,15 @@ class ProductInfo(BaseModel, extra=Extra.allow):
         self.state_ID = resource.data.ID
         self.event_time = event_time
         self.product_ID = _product.product_data.ID
-        self.activity = state.StateEnum.start_state
+        self.activity = StateEnum.start_state
         self.state_type = state_type
 
     def log_end_process(
         self,
         resource: resources.Resource,
-        _product: Product,
+        _product: "Product",
         event_time: float,
-        state_type: state.StateTypeEnum,
+        state_type: StateTypeEnum,
     ) -> None:
         """
         Logs the end of a process.
@@ -135,7 +131,7 @@ class ProductInfo(BaseModel, extra=Extra.allow):
         self.state_ID = resource.data.ID
         self.event_time = event_time
         self.product_ID = _product.product_data.ID
-        self.activity = state.StateEnum.end_state
+        self.activity = StateEnum.end_state
         self.state_type = state_type
 
 Locatable= Union[resources.Resource, node.Node, source.Source, sink.Sink]
@@ -155,17 +151,16 @@ class Product(BaseModel):
     env: sim.Environment
     product_data: product_data.ProductData
     process_model: proces_models.ProcessModel
-    transport_process: Union[process.LinkTransportProcess, process.TransportProcess, process.RequiredCapabilityProcess]
+    transport_process: Union[process.TransportProcess, process.RequiredCapabilityProcess, process.LinkTransportProcess]
     product_router: router.Router
 
     next_prodution_process: Optional[process.PROCESS_UNION] = Field(default=None, init=False)
     process: events.Process = Field(default=None, init=False)
     current_locatable: Locatable = Field(default=None, init=False)
     finished_process: events.Event = Field(default=None, init=False)
-    product_info: ProductInfo = ProductInfo()
+    product_info: ProductInfo = Field(default_factory=ProductInfo, init=False)
 
-    class Config:
-        arbitrary_types_allowed = True
+    model_config=ConfigDict(arbitrary_types_allowed=True)
 
     def update_location(self, resource: Locatable):
         """
@@ -197,11 +192,27 @@ class Product(BaseModel):
                 yield self.env.process(auxiliary_request_1.auxiliary.get_auxiliary(auxiliary_request))
 
         while self.next_prodution_process:
-            production_request, transport_request = yield self.env.process(self.product_router.route_product(self))
+            while True:
+                production_request = yield self.env.process(self.product_router.route_product_to_production_resource(self))
+                if not production_request:
+                    yield self.env.timeout(0)
+                    continue
+                break
+            while True:
+                transport_request = yield self.env.process(self.product_router.route_transport_resource_for_product(self, production_request))
+                if not transport_request:
+                    yield self.env.timeout(0)
+                    continue
+                break
             yield self.env.process(self.request_process(transport_request))
             yield self.env.process(self.request_process(production_request))
             self.set_next_production_process()
-        transport_to_sink_request = yield self.env.process(self.product_router.route_product_to_sink(self))
+        while True:
+            transport_to_sink_request = yield self.env.process(self.product_router.route_product_to_sink(self))
+            if not transport_to_sink_request:
+                yield self.env.timeout(0)
+                continue
+            break
         yield self.env.process(self.request_process(transport_to_sink_request))
 
         if self.product_data.auxiliaries:
@@ -258,9 +269,9 @@ class Product(BaseModel):
         Requests the next production process of the product object from the next production resource by creating a request event and registering it at the environment.
         """
         if isinstance(processing_request, request.TransportResquest):
-            type_ = state.StateTypeEnum.transport
+            type_ = StateTypeEnum.transport
         else:
-            type_ = state.StateTypeEnum.production
+            type_ = StateTypeEnum.production
         logger.debug({"ID": self.product_data.ID, "sim_time": self.env.now, "resource": processing_request.resource.data.ID, "event": f"Request process {processing_request.process.process_data.ID} for {type_}"})
         self.env.request_process_of_resource(
             request=processing_request
