@@ -35,16 +35,18 @@ source1 = psx.Source(product1, arrival_model_1, [0, 0], ID="source_1")
 source2 = psx.Source(product2, arrival_model_2, [0, 0], ID="source_2")
 
 
-
 system = psx.ProductionSystem([machine, machine2, transport], [source1, source2], [sink1, sink2])
 model = system.to_model()
 from prodsys import runner
 runner_instance = runner.Runner(adapter=model)
 runner_instance.initialize_simulation()
+simulation_source = runner_instance.source_factory.sources[0]
+product_example_1 = runner_instance.product_factory.create_product(simulation_source.product_data, simulation_source.router)
+print(product_example_1.product_data.ID)
 system.run(1000)
 
 runner_instance = system.runner
 
 runner_instance.print_results()
 runner_instance.plot_results()
-runner_instance.save_results_as_csv("examples")
+# runner_instance.save_results_as_csv("examples")
