@@ -4,8 +4,8 @@ from prodsys.optimization.math_opt import run_mathematical_optimization
 BASE_CONFIGURATION_FILE_PATH = "examples/optimization/optimization_example/base_scenario.json"
 SCENARIO_FILE_PATH = "examples/optimization/optimization_example/scenario.json"
 SAVE_FOLDER = "data/math_results"
-OPTIMIZATION_TIME_PORTION = 1.0
-NUMBER_OF_SOLUTIONS = 2
+OPTIMIZATION_TIME_PORTION = 1
+NUMBER_OF_SOLUTIONS = 1
 ADJUSTED_NUMBER_OF_TRANSPORT_RESOURCES = 2
 NUMBER_OF_SEEDS = 2
 
