@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import contextlib
 import random
-from pydantic import BaseModel, Field
+from pydantic import BaseModel, ConfigDict, Field
 from typing import List, Optional
 
 import numpy as np
@@ -10,7 +10,7 @@ import time
 from functools import cached_property
 
 from prodsys.adapters import adapter
-from prodsys.simulation import sim
+from prodsys.simulation import sim, logger
 from prodsys.factories import (
     link_transport_process_updater,
     auxiliary_factory,
@@ -24,8 +24,10 @@ from prodsys.factories import (
     source_factory,
     node_factory,
 )
-from prodsys.simulation import logger
-from prodsys.util import post_processing, kpi_visualization, util
+
+from prodsys.util.post_processing import PostProcessor
+
+from prodsys.util import kpi_visualization, util
 from prodsys.models import performance_data
 
 VERBOSE = 1
@@ -66,7 +68,7 @@ def temp_seed(seed: int):
         random.setstate(p_state)
 
 
-class Runner(BaseModel):
+class Runner:
     """
     Class to represent the simulation runner. It allows to run the simulation based on a provided adapter.
 
@@ -88,29 +90,24 @@ class Runner(BaseModel):
         time_stamp (str): The time stamp of the simulation run.
         post_processor (post_processing.PostProcessor): The post processor to process the simulation results.
     """
-    adapter: adapter.ProductionSystemAdapter
-    env: sim.Environment = Field(
-        None, description="The environment to run the simulation in", init=False
-    )
-    time_model_factory: time_model_factory.TimeModelFactory = Field(
-        init=False, default=None
-    )
-    state_factory: state_factory.StateFactory = Field(init=False, default=None)
-    process_factory: process_factory.ProcessFactory = Field(init=False, default=None)
-    queue_factory: queue_factory.QueueFactory = Field(init=False, default=None)
-    storage_factory: queue_factory.StorageFactory = Field(init=False, default=None)
-    auxiliary_factory: auxiliary_factory.AuxiliaryFactory = Field(init=False, default=None)
-    resource_factory: resource_factory.ResourceFactory = Field(init=False, default=None)
-    node_factory: node_factory.NodeFactory = Field(init=False, default=None)
-    sink_factory: sink_factory.SinkFactory = Field(init=False, default=None)
-    source_factory: source_factory.SourceFactory = Field(init=False, default=None)
-    product_factory: product_factory.ProductFactory = Field(init=False, default=None)
-    event_logger: logger.Logger = Field(init=False, default=None)
-    time_stamp: str = Field(init=False, default="")
-    post_processor: post_processing.PostProcessor = Field(init=False, default=None)
 
-    class Config:
-        arbitrary_types_allowed = True
+    def __init__(self, adapter: adapter.ProductionSystemAdapter):
+        self.adapter = adapter
+        self.env = sim.Environment(seed=self.adapter.seed)
+        self.time_model_factory: time_model_factory.TimeModelFactory = None
+        self.state_factory: state_factory.StateFactory = None
+        self.process_factory: process_factory.ProcessFactory = None
+        self.queue_factory: queue_factory.QueueFactory = None
+        self.auxiliary_factory: auxiliary_factory.AuxiliaryFactory = None
+        self.resource_factory: resource_factory.ResourceFactory = None
+        self.node_factory: node_factory.NodeFactory = None
+        self.sink_factory: sink_factory.SinkFactory = None
+        self.source_factory: source_factory.SourceFactory = None
+        self.product_factory: product_factory.ProductFactory = None
+        self.event_logger: logger.Logger = None
+        self.time_stamp: str = ""
+        self.post_processor: PostProcessor = None
+
 
     def initialize_simulation(self):
         """
@@ -136,9 +133,6 @@ class Runner(BaseModel):
 
             self.queue_factory = queue_factory.QueueFactory(env=self.env)
             self.queue_factory.create_queues(self.adapter)
-            
-            self.storage_factory = queue_factory.StorageFactory(env=self.env)
-            self.storage_factory.create_storages(self.adapter)
 
             self.resource_factory = resource_factory.ResourceFactory(
                 env=self.env,
@@ -168,7 +162,7 @@ class Runner(BaseModel):
             self.auxiliary_factory = auxiliary_factory.AuxiliaryFactory(
                 env=self.env,
                 process_factory=self.process_factory,
-                storage_factory=self.storage_factory,
+                storage_factory=self.queue_factory,
                 resource_factory=self.resource_factory,
                 sink_factory= self.sink_factory
             )
@@ -233,7 +227,7 @@ class Runner(BaseModel):
         t_1 = time.perf_counter()
         self.time_stamp = time.strftime("%Y%m%d-%H%M%S")
 
-    def get_post_processor(self) -> post_processing.PostProcessor:
+    def get_post_processor(self) -> PostProcessor:
         """
         Returns the post processor to process the simulation results.
 
@@ -241,7 +235,7 @@ class Runner(BaseModel):
             post_processing.PostProcessor: The post processor to process the simulation results.
         """
         if not self.post_processor:
-            self.post_processor = post_processing.PostProcessor(df_raw=self.event_logger.get_data_as_dataframe())
+            self.post_processor = PostProcessor(df_raw=self.event_logger.get_data_as_dataframe())
         return self.post_processor
 
 
@@ -259,9 +253,24 @@ class Runner(BaseModel):
         p = self.get_post_processor()
         kpi_visualization.plot_throughput_time_over_time(p)
         kpi_visualization.plot_WIP(p)
-        kpi_visualization.plot_WIP_per_resource(p)
+        # kpi_visualization.plot_WIP_per_resource(p)
         kpi_visualization.plot_throughput_time_distribution(p)
         kpi_visualization.plot_time_per_state_of_resources(p)
+    
+
+    def plot_results_executive(self):
+        """
+        Plots the aggregated simulation results, comprising the throughput time over time, WIP over time, throughput time distribution and the time per state of the resources.
+        """
+        p = self.get_post_processor()
+        kpi_visualization.plot_boxplot_resource_utilization(p)
+        kpi_visualization.plot_line_balance_kpis(p)
+        kpi_visualization.plot_production_flow_rate_per_product(p)
+        transport_resource_ids = [resource_data.ID for resource_data in adapter.get_transport_resources(self.adapter)]
+        kpi_visualization.plot_transport_utilization_over_time(p, transport_resource_ids)
+        kpi_visualization.plot_util_WIP_resource(p)
+        kpi_visualization.plot_oee(p)
+
 
     def get_event_data_of_simulation(self) -> List[performance_data.Event]:
         """
@@ -315,7 +324,7 @@ class Runner(BaseModel):
         Returns:
             dict: The aggregated simulation results.
         """
-        p = post_processing.PostProcessor(df_raw=self.event_logger.get_data_as_dataframe())
+        p = PostProcessor(df_raw=self.event_logger.get_data_as_dataframe())
         return p.get_aggregated_data()
 
     def save_results_as_csv(self, save_folder="data"):
